+++ NSelene/Conditions/Count.cs
@@ -23,12 +23,12 @@ namespace NSelene
 
             public override string DescribeActual()
             {
-                return this.actualCount.ToString();
+                return "count = " + this.actualCount;
             }
 
             public override string DescribeExpected()
             {
-                return this.expectedCount.ToString();
+                return "count = " + this.expectedCount;
             }
         }